lab13/hw/src/app/settings.py
from pydantic import computed_field
from pydantic_settings import BaseSettings, SettingsConfigDict


class AppSettings(BaseSettings):
    model_config = SettingsConfigDict(env_file=".env", extra="ignore")

    tavily_api_key: str
    tavily_mcp_base_url: str = "https://mcp.tavily.com/mcp/?tavilyApiKey="

    openai_api_key: str
    openai_base_url: str = "http://localhost:8000/v1"

    openweather_mcp_host: str = "localhost"
    openweather_mcp_port: int = 8001

    tool_loop_limit: int = 10
    max_completion_tokens: int = 1000

    @computed_field
    @property
    def mcp_servers(self) -> dict[str, str]:
        return {
            "tavily": f"{self.tavily_mcp_base_url}{self.tavily_api_key}",
            "openweather": f"http://{self.openweather_mcp_host}:{self.openweather_mcp_port}/mcp",
        }


_settings = AppSettings()  # type:ignore


def get_settings() -> AppSettings:
    return _settings
